Put the astropy.wcs include headers under its own directory so they aren't confused with other headers at the top-level (notably sip.h from PyQt).
