+++ CHANGES.rst
@@ -30,7 +30,7 @@ New Features
 
   - Use ``astropy.utils.iers`` to provide ``delta_ut1_utc``, so that
     automatic calculation of UT1 becomes possible [#1145].
-  
+
   - Add ``datetime`` format which allows converting to and from standard
     library ``datetime.datetime`` objects.
 
@@ -89,7 +89,7 @@ New Features
   - Added Brightness Temperature (antenna gain) equivalency for conversion
     between :math:`T_B` and flux density.
 
-  - Added percent unit, and allowed any string containing just a number 
+  - Added percent unit, and allowed any string containing just a number
     to be interpreted as a scaled dimensionless unit.
 
 - ``astropy.utils``
@@ -97,14 +97,14 @@ New Features
   - Added ``astropy.utils.iers`` which allows reading in of IERS A or
     IERS B bulletins and interpolation in UT1-UTC.
 
-- ``astropy.extern.six`` 
+- ``astropy.extern.six``
 
-  - Added `six <https://pypi.python.org/pypi/six/>`_ for python2/python3 
+  - Added `six <https://pypi.python.org/pypi/six/>`_ for python2/python3
     compatibility
 
 - Astropy now uses the ERFA library instead of the IAU SOFA library for
-  fundamental time transformation routines.  
-  The ERFA library is derived, with permission, from the IAU SOFA library but 
+  fundamental time transformation routines.
+  The ERFA library is derived, with permission, from the IAU SOFA library but
   is distributed under a BSD license. See ``license/ERFA.rst`` for details.
 
 - ``astropy.logger``
@@ -275,6 +275,17 @@ API Changes
   - The exception ``astropy.units.UnitsException`` has been renamed to
     ``astropy.units.UnitsError`` to be more consistent with the naming
     of built-in Python exceptions.
+
+- ``astropy.wcs``
+
+  - For those including the ``astropy.wcs`` C headers in their project, they
+    should now include it as:
+
+       #include "astropy_wcs/astropy_wcs_api.h"
+
+    instead of:
+
+       #include "astropy_wcs_api.h"
 
 Bug Fixes
 ^^^^^^^^^^
@@ -881,7 +892,7 @@ see the "What's New" section of the documentation for more details.
     astropy.cosmology classes. The current treatment assumes that neutrinos are
     massless. [#365]
 
-  - Add a WMAP9 object using the final (9-year) WMAP parameters from 
+  - Add a WMAP9 object using the final (9-year) WMAP parameters from
     Hinshaw et al. 2013. It has also been made the default cosmology
     [#629, #724].
