+++ astropy/wcs/include/astropy_wcs/isnan.h
