+++ astropy/wcs/include/astropy_wcs/pipeline.h
