+++ astropy/wcs/include/astropy_wcs/sip_wrap.h
