+++ astropy/wcs/include/astropy_wcs/str_list_proxy.h
