+++ astropy/wcs/src/wcslib_wtbarr_wrap.c
@@ -5,7 +5,7 @@
 
 #define NO_IMPORT_ARRAY
 
-#include "wcslib_wtbarr_wrap.h"
+#include "astropy_wcs/wcslib_wtbarr_wrap.h"
 
 #include <wcs.h>
 
@@ -15,7 +15,7 @@
  docstrings are written in doc/docstrings.py, which are then converted
  by setup.py into docstrings.h, which we include here.
 */
-#include "docstrings.h"
+#include "astropy_wcs/docstrings.h"
 
 /***************************************************************************
  * Helper functions                                                        *
